CloudFunction/EventDriven_CF/Python_Script/requirements.txt
pdfplumber
pandas
google-cloud-storage
openpyxl
